refactor: rename folder name from data to constants
