+++ apps/web/app/(app)/(detail)/research/lab/page.tsx
@@ -1,6 +1,6 @@
 import { PageHeader } from '../../../../../components/page-header'
 import { LabCard } from '../../../../../components/lab-card'
-import { LABS } from '../../../../../data/labs'
+import { LABS } from '../../../../../constants/labs'
 
 export default function Page(): JSX.Element {
   return (
