+++ apps/web/components/club-card.tsx
@@ -3,7 +3,7 @@ import { Link } from 'next-view-transitions'
 
 import { Card, CardHeader, CardTitle } from '@aics/ui/components/ui/card'
 import { AspectRatio } from '@aics/ui/components/ui/aspect-ratio'
-import type { Club } from '../data/clubs'
+import type { Club } from '../constants/clubs'
 
 interface ClubCardProps {
   club: Club
