+++ apps/web/components/lab-card.tsx
@@ -3,7 +3,7 @@ import { Link } from 'next-view-transitions'
 
 import { AspectRatio } from '@aics/ui/components/ui/aspect-ratio'
 import { Card, CardHeader, CardTitle } from '@aics/ui/components/ui/card'
-import type { Lab } from '../data/labs'
+import type { Lab } from '../constants/labs'
 
 function LabCard({ lab }: { lab: Lab }) {
   return (
