+++ apps/web/constants/curriculums.ts
